Add files via upload
This is a rough concept for the division operation of the ALU. Division is the most complex arithmetic operation to implement. It requires multiple components (full subtractors, magnitude comparators, and other components we have not yet discussed in lecture).  In order to completely design the division operation on a logic gate level, the other arithmetic components must be designed first.